Actor: reuse Vault tests (#668)
The Vault now exports its tests via a sharable utility.

See f41519a for the changes made to the original tests to make them sharable.
