+++ apps/vault/contracts/test/mocks/DestinationMock.sol
@@ -2,9 +2,18 @@ pragma solidity 0.4.24;
 
 
 contract DestinationMock {
-    uint256 test;
+    bool public expensiveFallback;
+    uint256 public counter;
+
+    constructor(bool _expensiveFallback) public {
+        expensiveFallback = _expensiveFallback;
+    }
 
     function () external payable {
-        test = test + 1;
+        // If expensiveFallback is used, this breaks the 2300 gas stipend given by
+        // .send() and .transfer()
+        if (expensiveFallback) {
+            counter = counter + 1;
+        }
     }
 }
