+++ future-apps/agent/contracts/test/TestImports.sol
@@ -8,6 +8,9 @@ import "@aragon/os/contracts/kernel/KernelProxy.sol";
 
 import "@aragon/apps-shared-migrations/contracts/Migrations.sol";
 import "@aragon/test-helpers/contracts/EtherTokenConstantMock.sol";
+import "@aragon/test-helpers/contracts/TokenMock.sol";
+import "@aragon/test-helpers/contracts/TokenReturnFalseMock.sol";
+import "@aragon/test-helpers/contracts/TokenReturnMissingMock.sol";
 
 // You might think this file is a bit odd, but let me explain.
 // We only use some contracts in our tests, which means Truffle
